// src/lib/api/generated/model/contactSubmissionCreateSchemaAffiliation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export type ContactSubmissionCreateSchemaAffiliation = typeof ContactSubmissionCreateSchemaAffiliation[keyof typeof ContactSubmissionCreateSchemaAffiliation];


export const ContactSubmissionCreateSchemaAffiliation = {
  student: 'student',
  parent: 'parent',
  faculty: 'faculty',
  staff: 'staff',
  visitor: 'visitor',
  other: 'other',
} as const;