// src/lib/api/generated/model/contactSubmissionCreateSchemaSubmissionType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export type ContactSubmissionCreateSchemaSubmissionType = typeof ContactSubmissionCreateSchemaSubmissionType[keyof typeof ContactSubmissionCreateSchemaSubmissionType];


export const ContactSubmissionCreateSchemaSubmissionType = {
  inquiry: 'inquiry',
  suggestion: 'suggestion',
  feedback: 'feedback',
  concern: 'concern',
  other: 'other',
} as const;